% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulationStudy}
\alias{simulationStudy}
\title{Execute numerical simulation}
\usage{
simulationStudy(model, interval, resEnh, anisotropy, est64, pattern)
}
\arguments{
\item{model}{type of model}

\item{interval}{The interval of current electrodes}

\item{resEnh}{Do resolution enhancement}

\item{anisotropy}{Consider the conductive anisotropy}

\item{est64}{shape parameter is 64 or 32}
}
\value{
normaly zero
}
\description{
Execute numerical simulation
}
